# plumbing: Optimise memory consumption for filesystem storage