@pjbgf 2023-07-08
Rebased and fixed issue with `ParserSuite.TestThinPack`.